kube-aggregator: split openapi tests
